Rename api dir to application
