+++ src/server/application/lib/initialize_contract.ts
