+++ src/server/application/lib/static_server.ts
