+++ src/server/index.ts
@@ -1,4 +1,4 @@
-import { bootApi } from './api'
+import { bootApi } from './application'
 
 const {
   API_PORT,
